\name{iterLap-internal}
\title{iterLap package internal functions}
\alias{getGrid}
\alias{getStartVal}
\alias{iterLap.control}
\alias{LSfitgeq0}
\alias{opt}
\alias{nlm.control}
\alias{samplingMixDist}
\alias{summary.mixDist}
\alias{print.mixDist}
\alias{print.summary.mixDist}
\alias{checkStop}
\alias{calcDist}
\alias{extractMoments}
\alias{dmvnormAdaptedC}
\alias{dmvtAdaptedC}
\alias{nlopt}
\alias{calcHess}

\description{
  These functions are not meant to be called by the user.
}

\keyword{ internal }
